% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrtruncnorm.R
\name{altrtruncnorm}
\alias{altrtruncnorm}
\title{Generate truncated random normal variables where values outside range are assigned to range limit}
\usage{
altrtruncnorm(n = 100, a = 0, b = 1, mean = 0, sd = 1)
}
\arguments{
\item{n}{the number of variables to be generated.}

\item{a}{the lower limit of range, positive numeric or positive numeric vector.}

\item{b}{the upper limit of range, positive numeric or positive numeric vector.}

\item{mean}{the mean of underlying normal distribution prior to truncation, positive numeric or positive numeric vector.}

\item{sd}{the standard deviation of underlying normal distribution prior to truncation, positive numeric or positive numeric vector.}
}
\description{
Note that this function is different from the one in the package truncnorm.
}
\details{
Updated 2018-08-14
}
\examples{
altrtruncnorm()
}
\keyword{misc}
