% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeasModstochApp.R
\name{SeasModstochApp}
\alias{SeasModstochApp}
\title{Generate an interactive Shiny interface for one simulation}
\usage{
SeasModstochApp(pHSinit = 0.8, Kx = 100, betax = 0.02,
  wxtnormm = 0.8, wxtnormsd = 0.3, hx = 1, mxtnormm = 1,
  mxtnormsd = 0.1, axtnormm = 1, axtnormsd = 0.1, rx = 0.1,
  zxtnormm = 1, zxtnormsd = 0.1, gx = 4, cx = 0.9, phix = 0,
  nseasons = 10, HPcut = 0.5, pHScut = 0.5, maY = 100, miY = 0,
  thetax = 0.2, Ex = 0)
}
\arguments{
\item{pHSinit}{initial proportion of healthy seed}

\item{Kx}{total number of plants}

\item{betax}{maximum seasonal transmission rate}

\item{wxtnormm}{environmental effect on transmission rate (mean of underlying normal distribution prior to truncation)}

\item{wxtnormsd}{environmental effect on transmission rate (standard deviation of underlying normal distribution prior to truncation)}

\item{hx}{host effect on transmission rate}

\item{mxtnormm}{vector management effect on transmission rate (mean of underlying normal distribution prior to truncation)}

\item{mxtnormsd}{vector management effect on transmission rate (standard deviation of underlying normal distribution prior to truncation)}

\item{axtnormm}{roguing effect in terms of decreased DP (mean of underlying normal distribution prior to truncation)}

\item{axtnormsd}{roguing effect in terms of decreased DP (standard deviation of underlying normal distribution prior to truncation)}

\item{rx}{reversion rate}

\item{zxtnormm}{proportional selection against diseased plants (mean of underlying normal distribution prior to truncation)}

\item{zxtnormsd}{proportional selection against diseased plants (standard deviation of underlying normal distribution prior to truncation)}

\item{gx}{seed production rate in healthy plants}

\item{cx}{proportional seed production rate in diseased plants}

\item{phix}{proportion clean seed purchased}

\item{nseasons}{number of seasons}

\item{HPcut}{proportion healthy plant number cutoff}

\item{pHScut}{proportion healthy seed cutoff}

\item{maY}{maximum attainable yield, end of season, in the absence of disease}

\item{miY}{minimum yield when all plants are diseased (useable yield despite disease)}

\item{thetax}{rate of decline of Yld with increasing disease incidence}

\item{Ex}{amount of external inoculum around field}
}
\description{
This function simulates one parameter combination across nseasons once
}
\details{
Updated 2018-08-23
}
\examples{
SeasModstochApp()
}
\keyword{health}
\keyword{seed}
